requirements.txt
opencv-python
numpy
pillow
SpeechRecognition
PyAudio
google-genai
python-dotenv
pymycobot
mediapipe
pybullet
ultralytics